First steps in autogenerating bindings with ffi_gen gem
`ffi_gen` is dark voodoo magic that uses clang/llvm to parse the header
and library files to autogenerate the Ruby code for the FFI bindings.
This is much more sustainable than hand curating the bindings and have
to work out the right incantation every time a method gets added or
updated with changes in the aubio source.