Breaking change: Aubio.beats is now an Array, not an enum
This change brings in several improvements to the Aubio.beats method to
make it easier to use with Sonic Pi. The update to Aubio v0.4.4 gives us
the ability to calculate the total number of samples in a file easily.
This can then be used to calculate the relative start and end points,
scaled between 0 and 1 which matches the Sonic Pi sample API

The enumerable (for large sound files) remains available as enum_beats

A new minioi param has been introduced to allow filtering of beats that
are too close together. This is set to a default of 0.15 seconds.